Move and remove some code.
